refactor: isolate FilaMan code for overlay reuse
Shrinks the FilaMan fork delta so it can be re-applied to a fresh Fluidd
checkout as a file overlay plus a small patch set.

Removed:
- widgets/filaman/SpoolSelectionDialog.vue (877 lines, no importer — the
  app-wide dialog in App.vue is widgets/spoolman/SpoolSelectionDialog.vue)
- unused serverFilamanProxyGetInfo and setExtruderSpool mutation
- cosmetic spoolman -> spoolTracking renames in FileSystem.vue,
  PrinterStatusCard.vue, MmuEditGateMapDialog.vue and
  AfcCardUnitLaneBody.vue (no behaviour change, pure merge friction)
- the serverSpoolmanPostSpoolId params tweak, a no-op since
  JSON.stringify drops undefined values

Extracted into overlay-owned modules:
- api/filamanActions.ts, util/filaman-spool-mapper.ts,
  store/spoolman/filamanActions.ts, typings/moonraker.filaman.d.ts
- store/spoolman/proxyResponse.ts holds the shared proxy helper so
  actions.ts and filamanActions.ts do not form an import cycle

FilaMan now registers its own MOONRAKER_COMPONENTS dispatch ahead of
spoolman and commits a backend flag, so store/server/actions.ts stays
upstream instead of forwarding the server info payload.
