+++ src/components/widgets/status/PrinterStatusCard.vue
@@ -161,10 +161,10 @@ export default class PrinterStatusCard extends Mixins(StateMixin, AfcMixin) {
       }
     }
 
-    const spoolTrackingAvailable: boolean = this.$typedGetters['spoolman/getAvailable']
+    const spoolmanSupported: boolean = this.$typedGetters['spoolman/getAvailable']
     const autoSpoolSelectionDialog: boolean = this.$typedState.config.uiSettings.spoolman.autoSpoolSelectionDialog
 
-    if (spoolTrackingAvailable && autoSpoolSelectionDialog) {
+    if (spoolmanSupported && autoSpoolSelectionDialog) {
       this.$typedCommit('spoolman/setDialogState', {
         show: true,
         filename
