+++ src/store/spoolman/state.ts
@@ -3,6 +3,7 @@ import type { SpoolmanState } from './types'
 export const defaultState = (): SpoolmanState => {
   return {
     info: null,
+    backend: 'spoolman',
     spools: [],
     activeSpool: null,
     activeSpoolsByExtruder: {},
